# mentor-marketplace-openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Marketplace Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for a real-time mentor marketplace backend.
    Includes authentication (OTP), user profiles, mentor discovery, booking lifecycle,
    session OTP start/end, and payment operations.
servers:
  - url: https://api.mentor-marketplace.com/v1
    description: Production
  - url: https://staging-api.mentor-marketplace.com/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP to phone number
      description: Sends a one-time password to the given phone for login/signup.
      security: []
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and issue tokens
      security: []
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /mentors/me/profile:
    get:
      tags: [Users]
      summary: Get mentor profile for current user
      operationId: getMyMentorProfile
      responses:
        '200':
          description: Mentor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Create or update mentor profile
      operationId: upsertMyMentorProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMentorProfileRequest'
      responses:
        '200':
          description: Mentor profile saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /mentors/search:
    get:
      tags: [Mentors]
      summary: Search mentors
      description: Returns mentors filtered by category/subcategory/availability/location and sorted by relevance.
      operationId: searchMentors
      parameters:
        - $ref: '#/components/parameters/CategoryId'
        - $ref: '#/components/parameters/SubcategoryId'
        - name: language
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 32
          required: false
        - name: availableOnly
          in: query
          schema:
            type: boolean
            default: true
          required: false
        - name: startTime
          in: query
          description: Requested start time for availability matching.
          schema:
            type: string
            format: date-time
          required: false
        - name: durationMinutes
          in: query
          description: Required duration to match availability.
          schema:
            type: integer
            minimum: 15
            maximum: 240
          required: false
        - name: locationCity
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 64
          required: false
        - name: locationCountry
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
            pattern: '^[A-Z]{2}$'
          required: false
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mentors/{mentorId}:
    get:
      tags: [Mentors]
      summary: Get mentor detail
      operationId: getMentorById
      parameters:
        - $ref: '#/components/parameters/MentorId'
      responses:
        '200':
          description: Mentor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings:
    post:
      tags: [Bookings]
      summary: Create booking (instant or scheduled)
      operationId: createBooking
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking details
      operationId: getBookingById
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{bookingId}/session-start-otp:
    post:
      tags: [Bookings]
      summary: Verify session start OTP and start session
      description: Session starts once required OTP validations complete (policy controlled server-side).
      operationId: startSessionWithOtp
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionOtpRequest'
      responses:
        '200':
          description: OTP accepted; session state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /bookings/{bookingId}/end:
    post:
      tags: [Bookings]
      summary: End session
      operationId: endSession
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndSessionRequest'
      responses:
        '200':
          description: Session ended and billing finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CategoryId:
      name: categoryId
      in: query
      required: false
      schema:
        type: string
        format: uuid
    SubcategoryId:
      name: subcategoryId
      in: query
      required: false
      schema:
        type: string
        format: uuid
    MentorId:
      name: mentorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: Request payload is invalid.
            requestId:
              type: string
              example: req_1A2B3C4D
...
    SendOtpRequest:
      type: object
      additionalProperties: false
      required: [phoneNumber, countryCode, purpose]
      properties:
        phoneNumber:
          type: string
          pattern: '^[0-9]{6,15}$'
          example: '9876543210'
        countryCode:
          type: string
          pattern: '^\+[1-9][0-9]{0,3}$'
          example: '+91'
        purpose:
          type: string
          enum: [login]
        deviceId:
          type: string
          maxLength: 128
    SendOtpResponse:
      type: object
      additionalProperties: false
      required: [otpRequestId, expiresAt, resendAfterSeconds, deliveryHint]
      properties:
        otpRequestId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        resendAfterSeconds:
          type: integer
          minimum: 0
        deliveryHint:
          type: string
          enum: [sms, dev_log, none]
          description: >
            sms: Twilio is configured; code should arrive by SMS.
            dev_log: OTP_LOG_CODE is on; code is printed in API logs.
            none: no SMS and no log mode; configure Twilio or OTP_LOG_CODE for local dev.
    VerifyOtpRequest:
      type: object
      additionalProperties: false
      required: [otpRequestId, otpCode]
      properties:
        otpRequestId:
          type: string
          format: uuid
        otpCode:
          type: string
          pattern: '^[0-9]{4,8}$'
        deviceId:
          type: string
          maxLength: 128
    VerifyOtpResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          minimum: 60
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, phoneNumber, roles]
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
        roles:
          type: array
          minItems: 1
          items:
            type: string
            enum: [mentee, mentor]
...
    UpsertMentorProfileRequest:
      type: object
      additionalProperties: false
      required: [bio, expertise, pricingPackages, availability, modalities, location]
      properties:
        bio:
          type: string
          minLength: 20
          maxLength: 2000
        expertise:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ExpertiseInput'
        modalities:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [video, voice, chat]
        pricingPackages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PricingPackageInput'
        availability:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AvailabilityInput'
        location:
          $ref: '#/components/schemas/LocationInput'
    MentorProfileResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/MentorProfile'
    MentorProfile:
      type: object
      additionalProperties: false
      required:
        [mentorId, bio, expertise, modalities, pricingPackages, availability, location, rating, totalSessions, isAvailableNow]
      properties:
        mentorId:
          type: string
          format: uuid
        bio:
          type: string
        expertise:
          type: array
          items:
            $ref: '#/components/schemas/Expertise'
        modalities:
          type: array
          items:
            type: string
            enum: [video, voice, chat]
        pricingPackages:
          type: array
          items:
            $ref: '#/components/schemas/PricingPackage'
        availability:
          type: array
          items:
            $ref: '#/components/schemas/Availability'
        location:
          $ref: '#/components/schemas/Location'
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        totalSessions:
          type: integer
          minimum: 0
        isAvailableNow:
          type: boolean
    ExpertiseInput:
      type: object
      additionalProperties: false
      required: [categoryId, subcategoryId, language]
      properties:
        categoryId:
          type: string
          format: uuid
        subcategoryId:
          type: string
          format: uuid
        language:
          type: string
          minLength: 2
          maxLength: 32
...
    PricingPackageInput:
      type: object
      additionalProperties: false
      required: [durationMinutes, priceAmount, currency]
      properties:
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 240
        priceAmount:
          type: integer
          minimum: 100
          description: Amount in minor unit (paise for INR).
        currency:
          type: string
          enum: [INR]
...
    MentorSearchResponse:
      type: object
      additionalProperties: false
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MentorSearchItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    MentorDetailResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/MentorSearchItem'
            - type: object
              additionalProperties: false
              required: [bio, modalities, availability]
              properties:
                bio:
                  type: string
                modalities:
                  type: array
                  items:
                    type: string
                    enum: [video, voice, chat]
                availability:
                  type: array
                  items:
                    $ref: '#/components/schemas/Availability'
    CreateBookingRequest:
      type: object
      additionalProperties: false
      required: [mentorId, packageId, bookingType, sessionMode]
      properties:
        mentorId:
          type: string
          format: uuid
        packageId:
          type: string
          format: uuid
        bookingType:
          type: string
          enum: [instant, scheduled]
        scheduledStartTime:
          type: string
          format: date-time
          description: Required when bookingType=scheduled.
        sessionMode:
          type: string
          enum: [video, voice, chat]
        notes:
          type: string
          maxLength: 500
    BookingResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Booking'
...
    StartSessionOtpRequest:
      type: object
      additionalProperties: false
      required: [otpCode, actorRole]
      properties:
        otpCode:
          type: string
          pattern: '^[0-9]{4,8}$'
        actorRole:
          type: string
          enum: [mentor, mentee]
    SessionStartResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [bookingId, status, otpValidatedBy, sessionStarted]
          properties:
            bookingId:
              type: string
              format: uuid
            status:
              type: string
              enum: [confirmed, started]
            otpValidatedBy:
              type: array
              items:
                type: string
                enum: [mentor, mentee]
            sessionStarted:
              type: boolean
            startedAt:
              type: string
              format: date-time
    EndSessionRequest:
      type: object
      additionalProperties: false
      required: [endedBy]
      properties:
        endedBy:
          type: string
          enum: [mentor, mentee, system]
        reason:
          type: string
          maxLength: 128
    EndSessionResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [bookingId, status, endedAt, billedAmount, currency]
          properties:
            bookingId:
              type: string
              format: uuid
            status:
              type: string
              enum: [completed, cancelled]
            endedAt:
              type: string
              format: date-time
            billedAmount:
              type: integer
              minimum: 0
              description: Amount in minor unit.
            currency:
              type: string
              enum: [INR]
...
    Location:
      allOf:
        - $ref: '#/components/schemas/LocationInput'
        - type: object
          additionalProperties: false
          properties:
            formattedAddress:
              type: string
...